+++ out/server/server.js
@@ -257,6 +257,7 @@ class LanguageServer {
                     if (split.length == 2) {
                         if (split[0] == 'constants') { //For example 'constants.parameterTypes'
                             const key = split[1];
+                            items.length = 0;
                             this.ucsjsHandler.AddConstants(items, this.ucsjsHandler.ucsjsConstants[key], split[1]);
                         }
                     }
@@ -272,7 +273,7 @@ class LanguageServer {
             else if (this.languageId == 'javascript') {
                 if (this.ucsjsHandler.isObject(items, linePrefix))
                     return items;
-                if (this.ucsjsHandler.isCVAsmManaged(items, prefixWord))
+                if (this.ucsjsHandler.isCVManaged(items, prefixWord))
                     return items;
                 if (this.ucsjsHandler.isLibraryClassInstances(items, linePrefix))
                     return items;
@@ -494,6 +495,7 @@ class LanguageServer {
         this.connection.onNotification('updateJSReferences', (params) => {
             this.ucsjsHandler.classLibraries = params.classRefs;
             this.ucsjsHandler.CVAsmManagedReferences = params.CVAsmManagedRefs;
+            this.ucsjsHandler.CVShapeManagedReferences = params.CVShapeManagedRefs;
             //console.log(`Received data updated references for libraries`);
         });
     }