+++ src/client/client.ts
@@ -157,6 +157,7 @@ export class LanguageClientWrapper {
       
       dynamicData.classRefs = this.ScriptProvider.UCSJSLibRefParser.classReferences;
       dynamicData.CVAsmManagedRefs = this.ScriptProvider.UCSJSLibRefParser.CVAsmManagedReferences;
+      dynamicData.CVShapeManagedRefs = this.ScriptProvider.UCSJSLibRefParser.CVShapeManagedReferences;
 
       // Send notification once the client is ready
       this.client.start().then(() => {