+++ src/server/ucsjsLanguageHandler.ts
@@ -10,7 +10,7 @@ import {
    } from 'vscode-languageserver/node';
    
    import * as fs from 'fs';
-import { UCSJSSystemConstants, UCSJSSystemPropertie, UCSJSSystemFunction, UCSJSSystemData,UCSJSSystemMethod, UCSJSParameterDef, DynamicData, docClassRef, classElement, ElementParam, CVAsmManaged } from '../interfaces';
+import { UCSJSSystemConstants, UCSJSSystemProperty, UCSJSSystemFunction, UCSJSSystemData,UCSJSSystemMethod, UCSJSParameterDef, DynamicData, docClassRef, classElement, ElementParam, CVManaged, UCSJSObject } from '../interfaces';
 import * as CONSTANTS from '../constants';
 import { Position, Uri } from 'vscode';
 
@@ -20,17 +20,18 @@ export class ucsjsLanguageHandler {
     private languageId: string;
 
 
-    private ucsjsObjects: string[] = [];
+    private ucsjsObjects: UCSJSObject[] = [];
     public ucsjsConstants: UCSJSSystemConstants = {} as UCSJSSystemConstants;
-    private ucsjsProperties: UCSJSSystemPropertie[] = [];
+    private ucsjsProperties: UCSJSSystemProperty[] = [];
     private ucsjsFunctions: UCSJSSystemFunction[] = [];
     public ucsjsMethods: UCSJSSystemMethod[] = [];
     public dynamicData: DynamicData = {} as DynamicData;
 
     public classLibraries: docClassRef[] = [];
 
-    public CVAsmManagedReferences: CVAsmManaged[] = [];
+    public CVAsmManagedReferences: CVManaged[] = [];
 
+    public CVShapeManagedReferences: CVManaged[] = [];
 
     // private AssemblyTypes: string[] = [];
     // private parameterModTypes: string[] = [];
@@ -69,12 +70,12 @@ export class ucsjsLanguageHandler {
     AddObjects(items: CompletionItem[]) {
         this.ucsjsObjects.forEach(obj => {
           items.push({
-            label: obj,
+            label: obj.name,
             kind: CompletionItemKind.Keyword,
             //detail: `**${obj}**\n\n (CV object)`
             documentation: {
                 kind: 'markdown',
-                value: `**${obj}**\n\n (CVAsmManaged object)`
+                value: `**${obj}**\n\n (${obj.Type} object)`
               }
           });
         });
@@ -163,19 +164,21 @@ export class ucsjsLanguageHandler {
         let Result: string = '';
         
         parameterDef.forEach(param =>
-            Result += `Type: ${param.ParamName}\n\n Description: ${param.ParamValue}\n\n`
+            Result += `\n\nType: ${param.ParamName}\n\n Description: ${param.ParamValue}`
         );
 
         return Result;
     }  
 
-    AddMethods(items: CompletionItem[],parentObject? : string) {
+    AddMethods(items: CompletionItem[],parentObject? : string,type?: string) {
         this.ucsjsMethods.forEach(method => {
             const pObj = parentObject ? parentObject : ''; 
             const paramDefs = this.buildMethodParams(method.parameterDef); 
-            const paramDefStr = paramDefs != '' ? `\n- **Parameters**: \n\n ${paramDefs}` : '';
+            const paramDefStr = paramDefs != '' ? `\n **Parameters**: ${paramDefs}` : '';
 
-          if (!parentObject && !method.parentObject || method.parentObject.includes(pObj)) {
+          //console.log(type , method.objectType, parentObject, method.parentObject);
+
+          if (!parentObject && !method.parentObject || method.parentObject.includes(pObj) || type && method.objectType == type) {
             items.push({
               label: method.name,
               kind: CompletionItemKind.Method,
@@ -191,15 +194,20 @@ export class ucsjsLanguageHandler {
         });
     }
 
-    AddProperties(items: CompletionItem[],parentObject? : string) {
+    AddProperties(items: CompletionItem[],parentObject? : string,type?: string) {
         const pObj = parentObject ? parentObject : ''; 
-
+//console.log(parentObject, type);
         this.ucsjsProperties.forEach(prop => {
-            if (!parentObject && !prop.parentObject || prop.parentObject.includes(pObj)) {   
+            if (!parentObject && !prop.parentObject || parentObject && prop.parentObject.includes(pObj) || type && prop.objectType == type) {
+                 console.log(prop.name , type, prop.objectType , parentObject , '-', pObj , '-', prop.parentObject);
                 items.push({
                 label: prop.name,
                 kind: CompletionItemKind.Property,
-                detail: `**${prop.name}**\n\n (${prop.Type} type)`,
+                documentation: {
+                    kind: 'markdown',
+                    value: `**${prop.name}**\n\n (${prop.Type} type)`
+                }
+                //detail: `**${prop.name}**\n\n (${prop.Type} type)`,
                 //   documentation: {
                 //     kind: 'markdown',
                 //     value: `**${prop.name}**\n\n- **Description**: ${prop.description}\n- **Value**: ${prop.value}\n- **Example**: ${prop.example}\n- **ReturnType**: ${prop.returnType}$`
@@ -211,10 +219,11 @@ export class ucsjsLanguageHandler {
 
     isObject(items: CompletionItem[],lineText: string)  : boolean {
         for (const spObj of this.ucsjsObjects) {
-            const wordRegex = new RegExp(`${spObj}[^\\s]*$`, 'i');
+            const wordRegex = new RegExp(`${spObj.name}[^\\s]*$`, 'i');
             if (wordRegex.test(lineText)) {
-                this.AddProperties(items,spObj);
-                this.AddMethods(items,spObj);
+                //console.log(lineText , spObj.name, spObj.Type);
+                this.AddProperties(items,spObj.name);
+                this.AddMethods(items,spObj.name);
                 return true;
             } 
         }
@@ -222,18 +231,24 @@ export class ucsjsLanguageHandler {
         return false;
     }
 
-    isCVAsmManaged(items: CompletionItem[],linePrefix: string)  : boolean {
-        for (const CVAsmObj of this.CVAsmManagedReferences) {
-            //console.log(linePrefix , CVAsmObj.variableName);
-            //const wordRegex = new RegExp(`${CVAsmObj.objectName}[^\\s]*$`, 'i');
-            if (linePrefix == CVAsmObj.variableName) {
-                this.AddProperties(items,CVAsmObj.objectName);
-                this.AddMethods(items,CVAsmObj.objectName);
-                console.log(linePrefix , CVAsmObj.variableName, CVAsmObj.objectName);
-                return true;
-            } 
+    isCVManaged(items: CompletionItem[],linePrefix: string)  : boolean {
+        const findCVManObj = (list:CVManaged[]) => {
+         for (const CVManObj of list) {
+             //console.log(linePrefix , CVManObj.variableName);
+             //const wordRegex = new RegExp(`${CVAsmObj.objectName}[^\\s]*$`, 'i');
+             if (linePrefix == CVManObj.variableName.toUpperCase()) {
+                //console.log(linePrefix , CVManObj.variableName);
+                 this.AddProperties(items,CVManObj.objectName,CVManObj.type);
+                 this.AddMethods(items,CVManObj.objectName,CVManObj.type);
+                 //console.log(linePrefix , CVManObj.variableName, CVManObj.objectName);
+                 return true;
+             } 
+         }
+         return false;
         }
         
+        if (findCVManObj(this.CVAsmManagedReferences)) return true;
+        if (findCVManObj(this.CVShapeManagedReferences)) return true;
         return false;
     }
 
@@ -277,12 +292,12 @@ export class ucsjsLanguageHandler {
 
     getHoverWord(word: string,wordRange: Range,prefixWord: string) : Hover | undefined {
 
-        const object = this.ucsjsObjects.find(obj => obj === word);
+        const object = this.ucsjsObjects.find(obj => obj.name === word);
         if (object) {
             return {
             contents: {
                 kind: 'markdown',
-                value: `**${object}**\n\n (CVAsmManaged object)`
+                value: `**${object.name}**\n\n (${object.Type} object)`
             },
             range: wordRange // Optional: Highlight the word
             };
@@ -299,11 +314,18 @@ export class ucsjsLanguageHandler {
             };
         }
 
-        const varRefMatch = this.CVAsmManagedReferences.find(varRef => varRef.variableName == prefixWord);
+        const CVAsmManVarRefMatch = this.CVAsmManagedReferences.find(varRef => varRef.variableName.toUpperCase() == prefixWord);
+        const CVShapeManVarRefMatch = this.CVShapeManagedReferences.find(varRef => varRef.variableName.toUpperCase() == prefixWord);
+        //this.connection.console.log(`variableName "${this.CVShapeManagedReferences[0].variableName}" objectName "${this.CVShapeManagedReferences[0].objectName}"`);
+        //this.connection.console.log(`prefixWord "${prefixWord}" objectType "${CVShapeManVarRefMatch?.type}" variableName "${CVShapeManVarRefMatch?.variableName}"`);
 
         const property = this.ucsjsProperties.find(prop => prop.name === word);
         if (property) { // 
-            if (property.parentObject.includes(prefixWord) || varRefMatch && property.parentObject.includes(varRefMatch.objectName)) {
+            const parObjUpper = property.parentObject ? property.parentObject.map(o => o.toUpperCase()) : [];
+            if (parObjUpper.includes(prefixWord) 
+                || CVAsmManVarRefMatch && property.parentObject.includes(CVAsmManVarRefMatch.objectName)
+                || CVShapeManVarRefMatch && property.objectType == CVShapeManVarRefMatch.type
+            ) {
                 //this.connection.console.log(`prefixWord "${prefixWord}" parentObject "${property.parentObject}"`);
                 return {
                 contents: {
@@ -318,10 +340,15 @@ export class ucsjsLanguageHandler {
         const method = this.ucsjsMethods.find(method => method.name === word);
         if (method) {
             //this.connection.console.log(`prefixWord "${prefixWord}" parentObject "${method.parentObject}"`);
-            if (method.parentObject.includes(prefixWord) || varRefMatch && method.parentObject.includes(varRefMatch.objectName)) {
+            const parObjUpper = method.parentObject.map(o => o.toUpperCase());
+            //this.connection.console.log(`prefixWord "${prefixWord}" parentObject "${method.parentObject}" parObjUpper "${parObjUpper}" objectName "${CVAsmManVarRefMatch?.objectName}"`);
+            if (parObjUpper.includes(prefixWord) 
+                || CVAsmManVarRefMatch && method.parentObject.includes(CVAsmManVarRefMatch.objectName)
+                || CVShapeManVarRefMatch && method.objectType == CVShapeManVarRefMatch.type
+            ) {
                 const paramDefs = this.buildMethodParams(method.parameterDef); 
                 const paramDefStr = paramDefs != '' ? `\n- **Parameters**: \n\n- ${paramDefs}` : '';
-                //this.connection.console.log(`Hover parameter data type "${method.name}"`);
+                this.connection.console.log(`Hover parameter data type "${method.name}"`);
                 return {
                 contents: {
                     kind: 'markdown',
@@ -369,6 +396,19 @@ export class ucsjsLanguageHandler {
                     contents: {
                         kind: 'markdown',
                         value: `**${CVAsmObj.variableName}**\n\n (CVAsmManaged object "${CVAsmObj.objectName}")`
+                    },
+                    range: wordRange // Optional: Highlight the word
+                    };
+            }
+        }
+
+        for (const CVManObj of this.CVShapeManagedReferences) {
+            if (CVManObj.variableName == word) {
+
+                return {
+                    contents: {
+                        kind: 'markdown',
+                        value: `**${CVManObj.variableName}**\n\n (CVShapeManaged object "${CVManObj.objectName}")`
                     },
                     range: wordRange // Optional: Highlight the word
                     };